pyproject.toml
[tool.poetry]
name = "python-gql"
version = "0.0.3"
description = "Python schema-first and auto-generate class graphql server"
authors = ["ysun <sunyu418@gmail.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "gql" }
]

[tool.poetry.dependencies]
python = "^3.7"
graphql-core = "^3"
click = "^7"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
isort = "*"
flake8 = "*"
pre-commit = "*"

[tool.poetry.scripts]
gqlgen = 'gql.cli.main:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
